Move kanji word actions up
